+++ CHANGELOG.md
@@ -462,3 +462,4 @@
 * Updated Visual Studio Editor package to 2.0.16
 * Implemented backtick option for string sub literal strings
 * Initial resolution can now auto-set to native instead of fixed (behind CoreParams.SetNativeResolutionOnFirstRun flag)
+* Implemented PseudoExtensibleEnum groundwork, with tests and custom property drawer (this is one line, but it's huge!)